requirements.txt
# Web framework
Flask>=3.0,<4.0
flask-login>=0.6,<1.0
werkzeug>=3.0,<4.0
Flask-WTF>=1.2,<2.0
Flask-Limiter>=3.5,<4.0

# Environment & config
python-dotenv>=1.0,<2.0

# Document parsing
PyPDF2>=3.0,<4.0
python-docx>=1.1,<2.0

# Data validation
pydantic>=2.0,<3.0

# LLM
groq>=0.9,<1.0

# Database
pymongo>=4.7,<5.0

# Numerical (for cosine similarity)
numpy>=1.26,<3.0

# HF Inference API
huggingface-hub>=0.30,<1.0

# Date parsing
python-dateutil>=2.9,<3.0
